fix: fix unsafe type conversion from int to float
There are a few issues with the uniform random rng strategy previously.
The first is that the static_cast from int to float for RAND_MAX is
making gcc raise an unsafe conversion warning. The second is that this
min-max rescaling of `rand()` is well-known to be an unsound way to
generate uniform random numbers due to the non-uniformity of floating
point coverage. Much better to rely on the std library for these things.