feat: move from time-based events to tick-based
There were several bugs with the anti-teaming feature that this commit
aims to fix, in addition to migrating from time-based to tick-based
simulation for the player object.

First, this commit ensures that all relevant stateful information is
owned by the individual players instead of globals owned by the engine.
The nice side-effect is that this stateful information is now under the
GameState state tree, instead of being stored off-tree in the Engine.

Then, this commit takes advantage of the fixed FPS tick rate allowing
us to freely convert from 60 ticks -> 1 second. Now, all time-based
events (such as decays and anti-teaming detection) are based on a number
of ticks. This conversion opens up several performance improvements,
such as being able to use "long ticks" to defer certain behavior checks
to only occur every `N` ticks.

Finally, this commit changes the algorithm used to detect anti-teaming.
Now, we keep track of the tick number where a virus was eaten and store
this in a list of eaten_virus_ticks. To detect anti-teaming, we then
filter that list for only the ticks that occur in the last 1 minute
(aka last 60 * 60 ticks) and count the number of occurrences. That
number of occurrences, determines the amount of anti-teaming penalty
being applied. This strategy ensures that the penalty scales up *and*
down as time goes own (previous strategy only scaled up and impacted
all players jointly).